+++ app/Http/Requests/Admin/UpdateUserRequest.php
@@ -26,7 +26,7 @@ class UpdateUserRequest extends FormRequest
         return [
             'name' => ['required','string','min:2','max:255'],
             'email' => ['required','string','lowercase','email','max:255', Rule::unique(User::class)->ignore(request()->route('user')->id)],
-            'role_id' => ['required','numeric', Rule::exists(Role::class, 'id')],
+            'role_id' => ['required','integer', Rule::exists(Role::class, 'id')],
         ];
     }
 }
